Add files via upload
Description of this project: 
This project is simple and represents the creation of an ordinary three-dimensional array. When launching the program, the user is prompted to select one of two items: 1). Creating a three-dimensional array without sorting; 2). Creating a three-dimensional array using sorting. I used quick sort as the sorting. Because it is quite fast (on average) and works well with large arrays.