three_dimensional_array
This project is quite simple. It involves creating a basic three-dimensional array. After starting the program, the user can choose one of two options: 1). Create a three-dimensional array without sorting; 2).Create a three-dimensional array with sorting. For sorting, I used the fast sorting algorithm because it tends to be fast and efficient for large amounts of processing.